fix::> One test added

+++ contracts/utils/JsonParser.sol
@@ -10,10 +10,10 @@ library JsonParser {
         PRIMITIVE
     }
 
-    uint256 public constant RETURN_SUCCESS = 0;
-    uint256 public constant RETURN_ERROR_INVALID_JSON = 1;
-    uint256 public constant RETURN_ERROR_PART = 2;
-    uint256 public constant RETURN_ERROR_NO_MEM = 3;
+    uint8 public constant RETURN_SUCCESS = 0;
+    uint8 public constant RETURN_ERROR_INVALID_JSON = 1;
+    uint8 public constant RETURN_ERROR_PART = 2;
+    uint8 public constant RETURN_ERROR_NO_MEM = 3;
 
     struct Token {
         JsonType jsonType;
